pyproject.toml
[build-system]
requires = ["setuptools>=64", "wheel", "setuptools-scm[toml]>=8"]
build-backend = "setuptools.build_meta"

[project]
name = "plex-generate-previews"
dynamic = ["version"]
authors = [
    {name = "Steve Adams", email = "stevezau@gmail.com"},
]
description = "Generate video preview thumbnails for Plex Media Server"
readme = "README.md"
license = "MIT"
requires-python = ">=3.10"
classifiers = [
    "Development Status :: 4 - Beta",
    "Intended Audience :: End Users/Desktop",
    "Operating System :: POSIX :: Linux",
    "Operating System :: Microsoft :: Windows",
    "Operating System :: MacOS",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Programming Language :: Python :: 3.12",
    "Topic :: Multimedia :: Video",
    "Topic :: System :: Archiving",
]
keywords = ["plex", "media", "server", "video", "thumbnails", "previews", "ffmpeg", "gpu", "acceleration"]
dependencies = [
    "requests==2.33.0",
    # urllib3 >=2.6.3 covers CVE-2025-50181/50182 (redirect-handling),
    # CVE-2025-66418/66471 (chunked encoding), CVE-2026-21441
    # (proxy-auth). python>=3.10 is already required so dropping 1.x
    # support is safe.
    "urllib3>=2.6.3,<3.0.0",
    # python-dotenv >=1.2.2 covers CVE-2026-28684.
    "python-dotenv>=1.2.2,<2",
    "plexapi==4.17.1",
    "loguru==0.7.3",
    "rich>=13.9,<14",
    "pymediainfo==7.0.1",
    "flask>=3.0.0,<4",
    "werkzeug>=3.0.0,<3.1.7",  # 3.1.7 rejects underscores in Host header, breaks Docker hostnames (#196)
    "flask-wtf>=1.2.0,<2",
    "flask-socketio>=5.3.0,<6",
    "flask-cors>=6.0.0,<7",  # 6.0.0 patches CVE-2024-6866 (case-insensitive path match)
    "flask-limiter>=3.5.0,<4",
    "apscheduler>=3.10.0,<4",
    "gunicorn>=22.0.0,<24",
    "sqlalchemy>=2.0.0,<3",
    # Pillow is used by JellyfinTrickplayAdapter to pack frame JPGs into
    # 10x10 tile sheets (Jellyfin 10.9+ native trickplay format).
    # >=12.2.0 covers CVE-2026-25990 and CVE-2026-40192.
    "pillow>=12.2.0,<14",
]

[project.optional-dependencies]
nvidia = ["pynvml==11.6.0"]
amd = ["amdsmi==6.2.1"]
test = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "pytest-mock>=3.10.0",
    "pytest-xdist>=3.0.0",
    "pytest-timeout>=2.3.0",
    "pytest-playwright>=0.4.0",
    "playwright>=1.40.0",
    "flask-cors>=6.0.0,<7",
]
dev = [
    "plex-generate-previews[test]",
    "ruff>=0.6.0",
    "pre-commit>=3.6.0",
    "locust>=2.20.0",
    "httpie>=3.2.0",
]

[project.urls]
Homepage = "https://github.com/stevezau/plex_generate_vid_previews"
Repository = "https://github.com/stevezau/plex_generate_vid_previews"
Issues = "https://github.com/stevezau/plex_generate_vid_previews/issues"

[tool.setuptools.packages.find]
where = ["."]
include = ["media_preview_generator*"]

[tool.pytest.ini_options]
testpaths = ["tests"]
addopts = [
    "--cov=media_preview_generator",
    "--cov-report=term-missing",
    "--cov-fail-under=70",
    "--tb=short",
    "--strict-markers",
    "--disable-warnings",
    "-n", "auto",
    "--dist", "load",
    "-m", "not gpu and not e2e and not integration",
    "--timeout=30",
    "--timeout-method=thread",
]
python_files = ["test_*.py"]
python_classes = ["Test*"]
python_functions = ["test_*"]
markers = [
    "integration: marks tests as integration tests",
    "real_plex_server: tests that require a live Plex server (opt out of the global Plex mock)",
    "real_gpu_detection: tests that require real GPU detection (opt out of the global GPU mock)",
    "gpu: marks tests that require GPU hardware",
    "plex: marks tests that require Plex server connection",
    "slow: marks tests as slow running",
    "ci: marks tests that work in CI environments",
    "e2e: marks tests as end-to-end browser tests",
    "real_prewarm: opt out of the autouse _prewarm_caches no-op in conftest",
    "real_logging: opt out of the autouse setup_logging no-op in conftest",
    "real_job_async: opt out of the autouse _start_job_async no-op in conftest",
]

[tool.coverage.run]
source = ["media_preview_generator"]
omit = [
    "*/tests/*",
    "*/__pycache__/*",
    "*/venv/*",
    "*/test_env/*"
]

[tool.coverage.report]
precision = 2
show_missing = true
skip_covered = false

[tool.setuptools_scm]
write_to = "media_preview_generator/_version.py"
local_scheme = "no-local-version"
fallback_version = "0.0.0+unknown"

[tool.ruff]
line-length = 120
target-version = "py310"
extend-exclude = [
    "media_preview_generator/_version.py",
    ".venv",
    ".venv-test",
    "htmlcov",
    "tmp_migration_test",
]

[tool.ruff.lint]
select = [
    "E",   # pycodestyle errors
    "W",   # pycodestyle warnings
    "F",   # pyflakes
    "I",   # isort
    "B",   # flake8-bugbear
    "UP",  # pyupgrade (modernizes typing, f-strings, etc.)
]
ignore = [
    "E501",  # line-too-long handled by formatter
    "B008",  # function calls in argument defaults (Flask/Depends idiom)
]

[tool.ruff.lint.per-file-ignores]
"tests/*" = ["B011"]  # assert False is fine in tests
"media_preview_generator/gpu/__init__.py" = ["F401"]  # public-API re-export
"media_preview_generator/processing/__init__.py" = ["F401"]  # public-API re-export

[tool.ruff.format]
quote-style = "double"
indent-style = "space"
